Add a session ID to incoming netconf sessions
Netconf hello messages should return a session ID as part of the response
(rfc6241 section 8.1) to a hello request from a client. This change
adds a session ID to each session to be used for such hello requests.
The session ID is derived from a global session ID that is assigned
randomly at start time. The global session ID is randomized so that
if the netconf server is restarted there is little chance of outstanding
requests with the same session ID.

While testing this fix it was noted that at least one client did not wait
for an rpc_ok message to be returned after sending a close-session request.
This caused error messages indicating a communication error. Now error
messages are not generated if sending a rpc_ok message in response to a
close-session message.