Fix crash on shutdown
If a netconf rpc is active when the server is shut down, then depending
on timing, a crash can occur. The crash is due to the sessions and g_schema
being deleted before an active thread in the thread pool is completed.
An issue with the routine free_open_sessions has also been fixed.