Add support for XPATH 1.0
XPATH 1.0 support has been added as per RFC 6241 sections 7.7 and 8.9.
XPATH 1.0 functionality is provided by libxml2, which we already use
to provide XML support. To use the XPATH functions requires that the
namespace of the XML data being queried be determined.

A new set of XPATH tests have been added to test suite. There are a few
XPATH 1.0 features such as @, text, comment that are not used in our XML
data, but tests have been added to make sure that using such terms does
not crash the server.