Fixed an issue determining the qdepth for a query.
A minor issue determining the last child has been fixed. It was actually
ending when it reached a GNode with no data. This was then causing the
following "if" to fail.

Now the report-all is working a test needed fixing. A new test has been
added that gets defaults for the test model. As the initial part of the
test the /test/settings branch of the apteryx tree is deleted, leaving
an empty branch with defaults that should be returned.