Fix some query issues.
If a valid query with a filter fails to match any data, then the result
returned is a query for all the data in the Apteryx database. This should
only happen for a query without a valid filter.

If an XPATH path has redundant namespace information in a query, then
it needs to be removed before using it. This is because we are only
storing namespace information in the Apteryx database at the root for a
model and at augmentation points. For example a query such as
/oc-if:interfaces/oc-if:interface/ needs to be reduced to
/oc-if:interfaces/interface/ so the XPATH engine in libxml2 can process
the query result correctly.