Add support for XML proxy nodes in a query
XML proxy nodes allow a schema to branch to using a remote apteryx
database for a query. This change works with a companion change
in apteryx-xml/schema.c. A copy of sch_path_to_gnode in schema.c
has been imported into data.c and modified to handle the complexities
of xpath parsing.