Change the ERROR macro used in data.c to the DEBUG macro.
When the XML code was shifted from apteryx-xml schema.c to
apteryx-netconf data.c the definition of the ERROR macro used
by the XML code changed. In schema.c the ERROR macro causes
a debug message to be generated, but only if the debug flag
is set. In apteryx-netconf the ERROR macro generates a syslog
message at severity LOG_ERR.

This change downgrades the ERROR macros used in data.c to being
DEBUG macros.