Tree from XML should not have duplicate nodes
When adding subtrees to a GNode tree ensure that if the root
of the tree being added is already a sub-node of the node to
which the subtree is to be added, that we add the subtree under
the existing node (removing the top of the subtree to do so).

This is not something that can be explicitly tested, but add
some tests to at least ensure that the new code doesn't cause
any problems.