xml_to_gnode: list key processing improvements
Key can now be anywhere in the message. Missing key when editing
will result in error.

Previous behaviour was to expect key to be the first child node, and
for a missing key to result in a '*' being inserted as the index,
which is OK for get, but not for edit.