edit-config: merge with just list index should work
A merge request creating a new list entry with only the list index
present in the message should work. It was actually being filtered
out because a merge request for an existing list entry with no
other data is a null operation. Rather than check for the existence
of the entry (which we would have to do for creates, deletes and
replaces) just go ahead and add the index to the merge list, even if
it risks having a do-nothing request.